+++ lib/process_hub/handler/cluster_update.ex
@@ -242,7 +242,8 @@ defmodule ProcessHub.Handler.ClusterUpdate do
     defp distribute_processes(%__MODULE__{} = args) do
       children = ProcessRegistry.registry(args.hub_id)
 
-      replication_factor = RedundancyStrategy.replication_factor(args.redun_strategy)
+      replication_factor =
+        RedundancyStrategy.replication_factor(args.redun_strategy, args.new_hash_ring)
 
       removed_node_processes(children, args.removed_node)
       |> Enum.each(fn {_, {child_spec, _}} ->